Refactor import paths for services and data objects
