--- src/app/sync/sync.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class SyncService {
-
-  constructor() { }
-}
